@@ -32,50 +32,48 @@
     </developers>
 
     <properties>
-        <jdk.version>1.8</jdk.version>
+        <jdk.version>11</jdk.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <graph.core.version>1.1.24</graph.core.version>
-        <analysis.core.version>3.3.7</analysis.core.version>
-        <diagram.exporter.version>1.3.18</diagram.exporter.version>
-        <reaction.exporter.version>1.1.4</reaction.exporter.version>
-        <fireworks.exporter.version>1.1.8</fireworks.exporter.version>
 
-        <junit.version>4.13.1</junit.version>
-        <itext.version>7.1.5</itext.version>
-        <spring.test.version>4.3.11.RELEASE</spring.test.version>
+        <graph.core.version>1.1.25-SNAPSHOT</graph.core.version>
+        <analysis.core.version>3.4.0-SNAPSHOT</analysis.core.version>
+        <diagram.exporter.version>1.4.0-SNAPSHOT</diagram.exporter.version>
+        <reaction.exporter.version>1.2.0-SNAPSHOT</reaction.exporter.version>
+        <fireworks.exporter.version>1.2.0-SNAPSHOT</fireworks.exporter.version>
 
-        <maven.compiler.version>3.5.1</maven.compiler.version>
-        <maven.jar.version>2.6</maven.jar.version>
+        <junit.version>5.8.0-M1</junit.version>
+        <itext.version>7.1.15</itext.version>
+        <spring.test.version>5.3.6</spring.test.version>
 
+        <maven.compiler.version>3.8.1</maven.compiler.version>
+        <maven.jar.version>3.2.0</maven.jar.version>
+        <maven.resources.version>3.2.0</maven.resources.version>
+        <maven.assembly.version>3.3.0</maven.assembly.version>
+        <maven.surefire.version>3.0.0-M5</maven.surefire.version>
+        <maven.source.version>3.2.1</maven.source.version>
     </properties>
 
-
     <dependencies>
-
         <dependency>
             <groupId>org.reactome.server.graph</groupId>
             <artifactId>graph-core</artifactId>
             <version>${graph.core.version}</version>
         </dependency>
-
         <dependency>
             <groupId>org.reactome.server.tools</groupId>
             <artifactId>analysis-core</artifactId>
             <version>${analysis.core.version}</version>
         </dependency>
-
         <dependency>
             <groupId>org.reactome.server.tools</groupId>
             <artifactId>reaction-exporter</artifactId>
             <version>${reaction.exporter.version}</version>
         </dependency>
-
         <dependency>
             <groupId>org.reactome.server.tools</groupId>
             <artifactId>diagram-exporter</artifactId>
             <version>${diagram.exporter.version}</version>
         </dependency>
-
         <dependency>
             <groupId>org.reactome.server.tools</groupId>
             <artifactId>fireworks-exporter</artifactId>
