@@ -1,8 +1,10 @@
 package org.reactome.server.tools.event.exporter;
 
-import org.junit.AfterClass;
-import org.junit.Before;
-import org.junit.runner.RunWith;
+
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.Assumptions;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.reactome.server.graph.aop.LazyFetchAspect;
 import org.reactome.server.graph.config.Neo4jConfig;
 import org.reactome.server.graph.service.GeneralService;
@@ -10,12 +12,11 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
 
-import static org.junit.Assume.assumeTrue;
 
-@ContextConfiguration(classes = { Neo4jConfig.class })
-@RunWith(SpringJUnit4ClassRunner.class)
+@ContextConfiguration(classes = {Neo4jConfig.class})
+@ExtendWith(SpringExtension.class)
 public abstract class BaseTest {
 
     static final Logger logger = LoggerFactory.getLogger("testLogger");
@@ -29,18 +30,16 @@ public abstract class BaseTest {
     @Autowired
     protected LazyFetchAspect lazyFetchAspect;
 
-    @AfterClass
+    @AfterAll
     public static void tearDownClass() {
     }
 
-    @Before
+    @BeforeEach
     public void setUp() {
         if (!checkedOnce) {
             isFit = generalService.fitForService();
             checkedOnce = true;
         }
-        assumeTrue(isFit);
+        Assumptions.assumeTrue(isFit);
     }
-
-
 }
